refactor menuItem controller

+++ src/main/java/pl/jakubtworek/RestaurantManagementSystem/repository/MenuItemRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 @Repository
 
-public interface MenuItemRepository extends JpaRepository<MenuItem, Integer> {
+public interface MenuItemRepository extends JpaRepository<MenuItem, Long> {
     List<MenuItem> findByMenu(Menu menu);
 }
